+++ R/polmineR.R
@@ -10,7 +10,6 @@ NULL
 #' @importFrom data.table data.table setorderv dcast setnames setkeyv setcolorder as.data.table rbindlist setkey dcast.data.table
 #' @importFrom Matrix rowSums colSums
 #' @importFrom RcppCWB cqp_is_initialized cqp_initialize
-#' @importFrom R6 R6Class
 #' @importFrom RcppCWB cl_attribute_size cl_lexicon_size cl_cpos2struc cl_cpos2id cl_struc2cpos cl_id2str cl_struc2str
 #' @importFrom RcppCWB cl_id2str cl_struc2str cl_regex2id cl_str2id cl_cpos2str cl_id2freq cl_id2cpos cl_cpos2lbound cl_cpos2rbound
 #' @importFrom RcppCWB cqp_query cqp_dump_subcorpus
@@ -60,7 +59,6 @@ setOldClass("htmlwidget")
 #'   \item{\emph{polmineR.browse}:} {}
 #'   \item{\emph{polmineR.buttons}:} {}
 #'   \item{\emph{polmineR.specialChars}:} {}
-#'   \item{\emph{polmineR.templates}:} {}
 #'   \item{\emph{polmineR.cutoff}:} {}
 #'   \item{\emph{polmineR.corpus_registry}:} {The system corpus registry
 #'   directory defined by the environment variable CORPUS_REGISTRY before the
@@ -146,7 +144,7 @@ polmineR <- function(){
     v <- tryCatch(utils::packageVersion(pkg), error = function(e) NA)
     if (is.na(v)) {
       msg <- sprintf("Package '%s' needed to run the shiny app.", pkg)
-      if (interactive()) {
+      if (interactive()){
         message(sprintf("%s Would you like to install it?", msg))
         if (utils::menu(c("Yes", "No")) == 1) {
           utils::install.packages(pkg)