+++ R/templates.R
@@ -1,74 +1,44 @@
 #' @include S4classes.R
 NULL
 
-#' Get and set templates.
+#' Get template for reconstructing full text.
 #' 
-#' Templates are used to format the markdown/html output of partitions. Upon loading 
-#' the polmineR package, templates for corpora are loaded into the option 'polmineR.templates'.
+#' Templates are used to format the markdown/html output of partitions.
 #' 
-#' @param .Object object 
-#' @param verbose logical, whether to be verbose
-#' @param ... further parameters
+#' @param .Object A \code{corpus}, \code{subcorpus} or \code{partition} object,
+#'   or a length-one \code{character} vector with a corpus ID.
+#' @param warn A \code{logical} value, whether to issue a warning if template 
+#'   is not available.
+#' @param ... Further arguments to be defined.
 #' @exportMethod get_template
-#' @exportMethod set_template
 #' @rdname templates
 #' @aliases get_template,subcorpus-method
 setGeneric("get_template", function(.Object, ...) standardGeneric("get_template"))
 
 #' @rdname templates
-setMethod("get_template", "character", function(.Object){
-  if (.Object %in% names(getOption("polmineR.templates"))){
-    return( getOption("polmineR.templates")[[.Object]] )
+setMethod("get_template", "character", function(.Object, warn = TRUE){
+  get_template(corpus(.Object), warn = warn)
+})
+
+#' @rdname templates
+setMethod("get_template", "corpus", function(.Object, warn = TRUE){
+  filename <- file.path(.Object@data_dir, "template.json")
+  if (file.exists(filename)){
+    y <- jsonlite::fromJSON(txt = filename) 
+    if ("metadata" %in% names(y)) y[["metadata"]] <- unlist(y[["metadata"]])
+    return(y)
   } else {
-    warning("the template requested is not available")
+    warning(sprintf("No template available for corpus '%s'.", .Object@corpus))
     return(NULL)
   }
 })
 
 #' @rdname templates
-setMethod("get_template", "corpus", function(.Object){
-  get_template(.Object@corpus)
+setMethod("get_template", "partition", function(.Object, warn = TRUE){
+  get_template(as(.Object, "corpus"), warn = TRUE)
 })
 
 #' @rdname templates
-setMethod("get_template", "partition", function(.Object){
-  getOption("polmineR.templates")[[.Object@corpus]]
-})
-
-#' @rdname templates
-setMethod("get_template", "subcorpus", function(.Object){
-  getOption("polmineR.templates")[[.Object@corpus]]
-})
-
-
-#' @rdname templates
-setMethod("get_template", "missing", function(.Object){
-  return( names(getOption("polmineR.templates")) ) 
-})
-
-#' @rdname templates
-setGeneric("set_template", function(.Object, ... ) standardGeneric("set_template"))
-
-#' @rdname templates
-setMethod("set_template", "character", function(.Object){
-  stopifnot(.Object %in% .list_corpora())
-  template_list <- getOption("polmineR.templates")
-  filename <- file.path(registry_get_home(.Object), "template.json")
-  if (file.exists(filename)){
-    template_list[[.Object]] <- jsonlite::fromJSON(txt = filename) 
-    if ("metadata" %in% names(template_list[[.Object]])){
-      template_list[[.Object]][["metadata"]] <- unlist(template_list[[.Object]][["metadata"]])
-    }
-  }
-  options("polmineR.templates" = template_list)
-  invisible(template_list[[.Object]])
-})
-
-
-#' @rdname templates
-#' @importFrom jsonlite fromJSON
-setMethod("set_template", "missing", function(.Object, verbose = FALSE){
-  if (length(Sys.getenv("CORPUS_REGISTRY")) > 0){
-    for (x in .list_corpora()) set_template(x)
-  }
+setMethod("get_template", "subcorpus", function(.Object, warn = TRUE){
+  callNextMethod()
 })